Move processors to the 'processors' namespace to be consistent
