+++ jpo-conflictmonitor/src/main/java/us/dot/its/jpo/conflictmonitor/monitor/processors/timestamp_deltas/BaseTimestampDeltaNotificationProcessor.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.conflictmonitor.monitor.topologies.timestamp_delta;
+package us.dot.its.jpo.conflictmonitor.monitor.processors.timestamp_deltas;
 
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.processor.Cancellable;
